users controllers added, auth controllers done (register, login), auth middleware done (isAuth) and general updates

+++ server/src/users/infrastructure/database/mongo.controllers.ts
@@ -10,13 +10,15 @@ class MongoControllers {
 
     public async getAllUsers(): Promise<UserEntityWithoutSensitiveData[] | null> {
         this.connectWithEnv();
-        const listOfUsers: Array<UserEntity> = await UserModel.find({});
+        const listOfUsers: Array<UserEntity> = await UserModel.find({})
+        .select('id username firstName LastName email age -_id');
         return listOfUsers;
     }
 
     public async getUserById(id: string): Promise<UserEntityWithoutSensitiveData | null>{
         this.connectWithEnv();
-        const user: UserEntity | null = await UserModel.findOne({id});
+        const user: UserEntity | null = await UserModel.findOne({id})
+        .select('id username firstName LastName email age -_id');
         return user;
     }
 
@@ -28,13 +30,17 @@ class MongoControllers {
 
     public async modifyUser(id: string, user: UserEntity): Promise<UserEntityWithoutSensitiveData | null> {
         this.connectWithEnv();
-        const modifiedUser = await UserModel.findOneAndUpdate({id}, user);
+        const modifiedUser = await UserModel.findOneAndUpdate({id}, user)
+        .select('id username firstName LastName email age -_id');
         return modifiedUser;
     }
 
     public async deleteUser(id: string): Promise<string | null> {
         this.connectWithEnv();
-        const deletedUser = await UserModel.findOneAndDelete({id});
+        const userAlreadyExists = await UserModel.findOne({id});
+        if(userAlreadyExists === null) return null
+        const deletedUser = await UserModel.findOneAndDelete({id})
+        .select('id username firstName LastName email age -_id');
         return deletedUser?.id;
     }
 
@@ -47,7 +53,8 @@ class MongoControllers {
 
     public async getUserByEmail(email: string): Promise<UserEntity | null> {
         this.connectWithEnv();
-        const user: UserEntity | null = await UserModel.findOne({email});
+        const user: UserEntity | null = await UserModel.findOne({email})
+        .select('id username firstName LastName email password age -_id');
         return user;
     }
 }
